type: reformatting
body:detail screen unexpected error
